requirements.txt
# Core dependencies for automated math solver
# GitHub API integration
PyGithub>=2.1.1
requests>=2.31.0

# Mathematical computation and symbolic math
sympy>=1.12
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0

# OCR and image processing for mathematical content
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Natural Language Processing for math problem understanding
nltk>=3.8.0
spacy>=3.6.0

# Web scraping and HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Scheduling and task management
schedule>=1.2.0
APScheduler>=3.10.0

# Email functionality
email-validator>=2.0.0

# Configuration and environment management
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Logging and monitoring
loguru>=0.7.0

# Web framework for API endpoints
flask>=2.3.0
flask-cors>=4.0.0

# Data validation
pydantic>=2.3.0

# File handling and utilities
watchdog>=3.0.0

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Math expression parsing
py-expression-eval>=0.3.14

# Additional math libraries
mpmath>=1.3.0
